Move C API functionality into c_api module
Move C API functionality into `c_api` module to clean up `lib.rs`.
